Better keyboard support
Bootstrap Style
Follow BS patterns and stop using `.active` to style when using the
keyboard to navigate. When using the arrow keys, the list items gain
`:focus`.

This makes handles future keypresses more difficult because you have to
detect them on the list item (grandchildren of the main component).
Additionally, you have to stop the arrow keys on keydown (not keyup
because it's too late) from scrolling the list.

ESC Press
I tried to just "do the right thing" in all cases, but there are many
cases to consider: open/closed, characters typed, arrow-ing up/down,
`showOnFocus` true/false.

TAB Press
This one has few cases, but need to ensure the browser is allowed to
move focus to the next item, so respond to the event, but don't stop it,
allow the browser to move focus normally after closing the list.

Cases Considered:
* close the list on ESC and TAB
* reopen the list if arrow keys are typed after closing
* ESC after arrowing closes the dropdown
* ESC in the text box clears any typed text
* ESC in the text box when no text is present closes the dropdown

Details:
* use `v-on="$listeners"` to pass listeners through to grandchildren
* must use @keydown (not @keyup) to stop event propagation. @keyup is
too late