Fix IE click handling
The related target field that chrome/firefox populate on blur events does
not appear to be a standard according to the MDN. However they are a standard
on the `focusout` event. So by moving over to the `focusout` event we are able
to capture the related target field on all tested browsers.

fixes #48