Support custom screen reader text per item
If `screenReaderText` is provided on an item, that text will be read
instead of the default text. This can be helpful for providing more
context for custom searches. For example you might have two
possibilities for a search text of "Western Front":
1) "Western Front title search"
2) "Western Front full text search"

If you make a custom display, the "text" might only be "Western Front",
which being read multiple times without additional context would be
unhelpful to a screen reader based user.