Improve screen reader text support
We needed to pass the screen reader text down into the list then into
the item. To follow the established convention, we added a screen reader
serializer allowing users to declare where to find there screen reader
text.

In an effort to provide sensible defaults we default screen reader text
to the output of the serializer if it exists for an object or just the
text if we're working with an array of strings.